requirements.txt
numpy
scipy
pyyaml
fitsio
matplotlib
esutil
healpy
